+++ irodsadmin/management/commands/irodsorphans.py
@@ -11,34 +11,73 @@ from landingzones.models import LandingZone
 
 # Samplesheets dependency
 from samplesheets.models import Assay, Study
+from samplesheets.plugins import find_assay_plugin
+from samplesheets.rendering import SampleSheetTableBuilder
+from samplesheets.views import TRACK_HUBS_COLL, RESULTS_COLL, MISC_FILES_COLL
 
 
-def get_assay_collections():
+def get_assay_collections(assays, irods_backend):
     """Return a list of all assay collection names."""
-    return ['assay_{}'.format(a.sodar_uuid) for a in Assay.objects.all()]
+    return [irods_backend.get_path(a) for a in assays]
 
 
-def get_study_collections():
+def get_assay_subcollections(studies, irods_backend):
+    """Return a list of all assay row colletion names."""
+    tb = SampleSheetTableBuilder()
+
+    collections = []
+
+    for study in studies:
+        study_tables = tb.build_study_tables(study)
+        for assay in study.assays.all():
+            assay_table = study_tables['assays'][str(assay.sodar_uuid)]
+            assay_plugin = find_assay_plugin(
+                assay.measurement_type, assay.technology_type
+            )
+            assay_path = irods_backend.get_path(assay)
+            if assay_plugin:
+                for row in assay_table['table_data']:
+                    row_path = assay_plugin.get_row_path(
+                        row, assay_table, assay, assay_path
+                    )
+                    if row_path not in collections:
+                        collections.append(row_path)
+
+                shortcuts = assay_plugin.get_shortcuts(assay)
+
+                if shortcuts:
+                    for shortcut in shortcuts:
+                        collections.append(shortcut['path'])
+
+                # Add default expected subcollections of assay collection
+                collections.append(assay_path + '/' + TRACK_HUBS_COLL)
+                collections.append(assay_path + '/' + RESULTS_COLL)
+                collections.append(assay_path + '/' + MISC_FILES_COLL)
+
+    return collections
+
+
+def get_study_collections(studies, irods_backend):
     """Return a list of all study collection names."""
-    return ['study_{}'.format(s.sodar_uuid) for s in Study.objects.all()]
+    return [irods_backend.get_path(s) for s in studies]
 
 
-def get_zone_collections():
+def get_zone_collections(irods_backend):
     """
     Return a list of all landing zone collection names that are not MOVED or
     DELETED.
     """
     return [
-        l.title
-        for l in LandingZone.objects.filter(
+        irods_backend.get_path(lz)
+        for lz in LandingZone.objects.filter(
             ~(Q(status='MOVED') & Q(status='DELETED'))
         )
     ]
 
 
-def get_project_collections():
+def get_project_collections(irods_backend):
     """Return a list of all study collection names."""
-    return [str(p.sodar_uuid) for p in Project.objects.all()]
+    return [irods_backend.get_path(p) for p in Project.objects.all()]
 
 
 def is_zone(collection):
@@ -73,7 +112,7 @@ def is_project(collection):
     )
 
 
-def get_orphans(session, irods_backend, expected):
+def get_orphans(session, irods_backend, expected, assays):
     """
     Return a list of orphans in a given irods session that are not in a given
     list of expected collections.
@@ -87,9 +126,18 @@ def get_orphans(session, irods_backend, expected):
             or is_assay_or_study(collection)
             or is_project(collection)
         ):
-            if collection.name in expected:
-                continue
-            orphans.append(collection.path)
+            if collection.path not in expected:
+                orphans.append(collection.path)
+
+    for assay in assays:
+        if not find_assay_plugin(assay.measurement_type, assay.technology_type):
+            continue
+
+        for collection in irods_backend.get_child_colls_by_path(
+            irods_backend.get_path(assay)
+        ):
+            if collection.path not in expected:
+                orphans.append(collection.path)
 
     return orphans
 
@@ -102,11 +150,16 @@ class Command(BaseCommand):
     def handle(self, *args, **options):
         irods_backend = get_backend_api('omics_irods')
         session = irods_backend.get_session()
+        studies = list(Study.objects.all())
+        assays = list(Assay.objects.all())
         expected = (
-            *get_assay_collections(),
-            *get_study_collections(),
-            *get_zone_collections(),
-            *get_project_collections(),
+            *get_assay_collections(assays, irods_backend),
+            *get_study_collections(studies, irods_backend),
+            *get_zone_collections(irods_backend),
+            *get_project_collections(irods_backend),
+            *get_assay_subcollections(studies, irods_backend),
         )
-        orphans = get_orphans(session, irods_backend, expected)
-        self.stdout.write('\n'.join(orphans))
+
+        orphans = get_orphans(session, irods_backend, expected, assays)
+        if orphans:
+            self.stdout.write('\n'.join(orphans))