+++ samplesheets/views_ajax.py
@@ -1,5 +1,4 @@
 """Ajax API views for the samplesheets app"""
-import os
 
 from altamisa.constants import table_headers as th
 from datetime import datetime as dt
@@ -25,7 +24,6 @@ from samplesheets.models import (
     Protocol,
     Process,
     GenericMaterial,
-    IrodsAccessTicket,
 )
 from samplesheets.sheet_config import SheetConfigAPI
 from samplesheets.rendering import (
@@ -41,10 +39,6 @@ from samplesheets.views import (
     app_settings,
     APP_NAME,
     TARGET_ALTAMISA_VERSION,
-    RESULTS_COLL_ID,
-    RESULTS_COLL,
-    MISC_FILES_COLL_ID,
-    MISC_FILES_COLL,
     logger,
 )
 
@@ -453,9 +447,10 @@ class SampleSheetStudyTablesAjaxView(SODARBaseProjectAjaxView):
         return display_config
 
     def get(self, request, *args, **kwargs):
+        from samplesheets.plugins import get_irods_content
+
         timeline = get_backend_api('timeline_backend')
         irods_backend = get_backend_api('omics_irods', conn=False)
-        cache_backend = get_backend_api('sodar_cache')
         study = Study.objects.filter(sodar_uuid=self.kwargs['study']).first()
 
         if not study:
@@ -501,138 +496,8 @@ class SampleSheetStudyTablesAjaxView(SODARBaseProjectAjaxView):
             return Response(ret_data, status=200)
 
         # Get iRODS content if NOT editing and collections have been created
-        if not edit and inv.irods_status and irods_backend:
-            # Can't import at module root due to circular dependency
-            from .plugins import find_study_plugin
-            from .plugins import find_assay_plugin
-
-            # Get study plugin for shortcut data
-            study_plugin = find_study_plugin(inv.get_configuration())
-
-            if study_plugin:
-                shortcuts = study_plugin.get_shortcut_column(
-                    study, ret_data['tables']
-                )
-                ret_data['tables']['study']['shortcuts'] = shortcuts
-
-            # Get assay content if corresponding assay plugin exists
-            for a_uuid, a_data in ret_data['tables']['assays'].items():
-                assay = Assay.objects.filter(sodar_uuid=a_uuid).first()
-                assay_path = irods_backend.get_path(assay)
-                a_data['irods_paths'] = []
-
-                # Default shortcuts
-                a_data['shortcuts'] = [
-                    {
-                        'id': RESULTS_COLL_ID,
-                        'label': 'Results and Reports',
-                        'path': assay_path + '/' + RESULTS_COLL,
-                        'assay_plugin': False,
-                    },
-                    {
-                        'id': MISC_FILES_COLL_ID,
-                        'label': 'Misc Files',
-                        'path': assay_path + '/' + MISC_FILES_COLL,
-                        'assay_plugin': False,
-                    },
-                ]
-
-                assay_plugin = find_assay_plugin(
-                    assay.measurement_type, assay.technology_type
-                )
-
-                if assay_plugin:
-                    cache_item = cache_backend.get_cache_item(
-                        name='irods/rows/{}'.format(a_uuid),
-                        app_name=assay_plugin.app_name,
-                        project=assay.get_project(),
-                    )
-
-                    for row in a_data['table_data']:
-                        # Update assay links column
-                        path = assay_plugin.get_row_path(
-                            row, a_data, assay, assay_path
-                        )
-                        enabled = True
-
-                        # Set initial state to disabled by cached value
-                        if (
-                            cache_item
-                            and path in cache_item.data['paths']
-                            and (
-                                not cache_item.data['paths'][path]
-                                or cache_item.data['paths'][path] == 0
-                            )
-                        ):
-                            enabled = False
-
-                        a_data['irods_paths'].append(
-                            {'path': path, 'enabled': enabled}
-                        )
-                        # Update row links
-                        assay_plugin.update_row(row, a_data, assay)
-
-                    assay_shortcuts = assay_plugin.get_shortcuts(assay) or []
-
-                    # Add visual notification to all shortcuts coming from an assay plugin
-                    for a in assay_shortcuts:
-                        a['icon'] = 'fa-puzzle-piece'
-                        a['title'] = 'Defined in assay plugin'
-                        a['assay_plugin'] = True
-
-                    # Add extra table if available
-                    a_data['shortcuts'].extend(assay_shortcuts)
-
-                # Check assay shortcut cache and set initial enabled value
-                cache_item = cache_backend.get_cache_item(
-                    name='irods/shortcuts/assay/{}'.format(a_uuid),
-                    app_name=APP_NAME,
-                    project=assay.get_project(),
-                )
-
-                # Add track hub shortcuts
-                track_hubs = (
-                    cache_item
-                    and cache_item.data['shortcuts'].get('track_hubs')
-                ) or []
-
-                for i, track_hub in enumerate(track_hubs):
-                    tickets = IrodsAccessTicket.active_objects.filter(
-                        path=track_hub
-                    ) or IrodsAccessTicket.objects.filter(path=track_hub)
-                    ticket = tickets and tickets.first()
-                    a_data['shortcuts'].append(
-                        {
-                            'id': 'track_hub_%d' % i,
-                            'label': os.path.basename(track_hub),
-                            'icon': 'fa-road',
-                            'title': 'Track Hub',
-                            'assay_plugin': False,
-                            'path': track_hub,
-                            'extra_links': [
-                                {
-                                    'url': ticket.get_webdav_link(),
-                                    'icon': 'fa-ticket',
-                                    'id': 'ticket_access_%d' % i,
-                                    'class': 'sodar-irods-ticket-access-%d-btn'
-                                    % i,
-                                    'title': ' iRODS Access Ticket',
-                                    'enabled': ticket.is_active(),
-                                }
-                            ]
-                            if ticket
-                            else [],
-                        }
-                    )
-
-                for i in range(len(a_data['shortcuts'])):
-                    if cache_item:
-                        a_data['shortcuts'][i]['enabled'] = cache_item.data[
-                            'shortcuts'
-                        ].get(a_data['shortcuts'][i]['id'])
-
-                    else:
-                        a_data['shortcuts'][i]['enabled'] = True
+        if not edit:
+            ret_data = get_irods_content(inv, study, irods_backend, ret_data)
 
         # Get/build sheet config
         sheet_config = conf_api.get_sheet_config(inv)